+++ core/api/typescript/src/lib/enclaves/generic_api_container_client.ts
@@ -4,26 +4,22 @@
  */
 
 import { Result } from "neverthrow";
-import * as google_protobuf_empty_pb from "google-protobuf/google/protobuf/empty_pb";
 import {
     ExecCommandArgs,
     ExecCommandResponse,
     ExecuteModuleArgs,
-    ExecuteModuleResponse,
-    GetModuleInfoArgs,
-    GetModulesResponse,
-    GetServiceInfoArgs,
-    GetServiceInfoResponse,
+    ExecuteModuleResponse, GetModulesArgs,
+    GetModulesResponse, GetServicesArgs,
     GetServicesResponse,
     LoadModuleArgs,
     PauseServiceArgs,
     RegisterServiceArgs,
     RegisterServiceResponse,
-    RemoveServiceArgs,
+    RemoveServiceArgs, RemoveServiceResponse,
     RepartitionArgs,
     StartServiceArgs,
     StartServiceResponse, StoreFilesArtifactFromServiceArgs, StoreWebFilesArtifactArgs, StoreWebFilesArtifactResponse,
-    UnloadModuleArgs,
+    UnloadModuleArgs, UnloadModuleResponse,
     UnpauseServiceArgs,
     UploadFilesArtifactArgs,
     UploadFilesArtifactResponse,
@@ -35,17 +31,15 @@ import { EnclaveID } from "./enclave_context";
 export interface GenericApiContainerClient {
     getEnclaveId(): EnclaveID
     loadModule(loadModuleArgs: LoadModuleArgs): Promise<Result<null, Error>>
-    unloadModule(unloadModuleArgs: UnloadModuleArgs): Promise<Result<null,Error>>
-    getModuleInfo(getModuleInfoArgs: GetModuleInfoArgs): Promise<Result<null, Error>>
+    unloadModule(unloadModuleArgs: UnloadModuleArgs): Promise<Result<UnloadModuleResponse,Error>>
     registerService(registerServiceArgs: RegisterServiceArgs): Promise<Result<RegisterServiceResponse, Error>>
     startService(startServiceArgs: StartServiceArgs): Promise<Result<StartServiceResponse, Error>>
-    getServiceInfo(getServiceInfoArgs: GetServiceInfoArgs): Promise<Result<GetServiceInfoResponse, Error>>
-    removeService(args: RemoveServiceArgs): Promise<Result<null, Error>>
+    removeService(args: RemoveServiceArgs): Promise<Result<RemoveServiceResponse, Error>>
     repartitionNetwork(repartitionArgs: RepartitionArgs): Promise<Result<null, Error>>
     waitForHttpGetEndpointAvailability(availabilityArgs: WaitForHttpGetEndpointAvailabilityArgs): Promise<Result<null, Error>>
     waitForHttpPostEndpointAvailability(availabilityArgs: WaitForHttpPostEndpointAvailabilityArgs): Promise<Result<null, Error>>
-    getServices(emptyArg: google_protobuf_empty_pb.Empty): Promise<Result<GetServicesResponse, Error>>
-    getModules(emptyArg: google_protobuf_empty_pb.Empty): Promise<Result<GetModulesResponse, Error>>
+    getServices(getServicesArgs: GetServicesArgs): Promise<Result<GetServicesResponse, Error>>
+    getModules(getModulesArgs: GetModulesArgs): Promise<Result<GetModulesResponse, Error>>
     executeModule(executeModuleArgs: ExecuteModuleArgs): Promise<Result<ExecuteModuleResponse, Error>>
     execCommand(execCommandArgs: ExecCommandArgs): Promise<Result<ExecCommandResponse, Error>>
     pauseService(pauseServiceArgs: PauseServiceArgs): Promise<Result<null, Error>>